Lesson 2, Level 1, Task 1, updated with double divide.

+++ src/main/java/students/romans_bogomols/lesson_2/level_1/task_1/Task1.java
@@ -16,7 +16,7 @@ public class Task1 {
         int secondNumber = scanner.nextInt();
         int sumOfNumbers = firstNumber + secondNumber;
         int multipleOfNumbers = firstNumber * secondNumber;
-        int divideOfNumbers = firstNumber / secondNumber;
+        double divideOfNumbers = firstNumber / (double)secondNumber;
         int subtractionOfNumbers = firstNumber - secondNumber;
 
         System.out.println(firstNumber + " + " + secondNumber + " = " + sumOfNumbers);
